Initial wrapper around debian-installed system libfluent
FluentBit distributes their own debian packages which are slightly
misbehaved.  The package "fluent-bit" installs the binary at
    /opt/fluent-bit/bin/fluent-bit
and the lib at
    /lib/fluent-bit/libfluent-bit.so
and the package "fluent-bit-headers" installs to
    /opt/fluent-bit/include/
but lacks lots of headers fluent-bit.h depends on, so they aren't
usable.

This alien wrapper checks for the system libfluentbit at those
locations, and if found it retrieves the version from the fluentbit
command, and adds its own minimal header.

I have not yet figured out how to create a static library.