Fix up the 'share' installed recipe and include fluent-bit-minimal.h
The share-built recipe works now, although the headers are still
missing.  I am including the same minimal header for the share-built
version as for the system-installed version.  I decided to rename
the header to fluent-bit-minimal.h and install it alongside the
others, and also clone it overtop the fluent-bit.h until upstream
fixes their cmake config.